+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2021-10-07
+
+### Added
+
+- Added `BaseHelpCommand` to show command help
+
+### Changed
+
+- BaseStartup is now part of namespace `Zen.Host`
+- Json configuration is no longer added by default.
+
+### Removed
+
+- Removed dependency on Autofac
+- Removed dependency of Microsoft.Extensions.Configuration.Json
+- Removed `ConfigureContainer` function from BaseStartup
+
 ## [1.5.0] - 2021-09-27
 
 ### Added 
@@ -63,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added extension function for Autofac to add all commands from assembly to DI
 - Added Base Command containing execute and validate functions 
 
-[Unreleased]: https://github.com/WajahatAliAbid/zen-clifx-extensions/compare/1.4.0...HEAD
+[Unreleased]: https://github.com/WajahatAliAbid/zen-clifx-extensions/compare/1.6.0...HEAD
+[1.6.0]: https://github.com/WajahatAliAbid/zen-clifx-extensions/compare/1.5.0...1.6.0
 [1.5.0]: https://github.com/WajahatAliAbid/zen-clifx-extensions/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/WajahatAliAbid/zen-clifx-extensions/compare/1.3.1...1.4.0
 [1.3.1]: https://github.com/WajahatAliAbid/zen-clifx-extensions/compare/1.3.0...1.3.1